requirements.txt
mistralai>=0.0.1
python-dotenv>=1.0.0
